requirements.txt
PyYAML==5.4.1
Flask==2.0.1
marshmallow==3.12.1
azure-storage-blob==12.8.1
azure-identity==1.6.0